# src/NetShield.Web.Host/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "NetShield API",
    "description": "The NetShield API. Every path is versioned under /api/v1 and every error is an RFC 9457 problem document.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "ListDevices",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "descending",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceState"
            }
          },
          {
            "name": "vendor",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceVendor"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceRole"
            }
          },
          {
            "name": "criticality",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CriticalityTier"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceEnvironment"
            }
          },
          {
            "name": "site",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageOfDeviceSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "operationId": "CreateDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "operationId": "GetDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "operationId": "UpdateDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "operationId": "DeleteDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDeviceRequest": {
        "required": [
          "hostname",
          "primaryIpAddress"
        ],
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "primaryIpAddress": {
            "type": "string"
          },
          "vendor": {
            "$ref": "#/components/schemas/DeviceVendor"
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "osVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "serialNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/DeviceRole"
          },
          "criticality": {
            "$ref": "#/components/schemas/CriticalityTier"
          },
          "environment": {
            "$ref": "#/components/schemas/DeviceEnvironment"
          },
          "owner": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CriticalityTier": {
        "type": "integer"
      },
...
      "CursorPageOfDeviceSummary": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceSummary"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "DeviceDetail": {
        "required": [
          "id",
          "hostname",
          "primaryIpAddress",
          "vendor",
          "model",
          "osVersion",
          "serialNumber",
          "site",
          "role",
          "criticality",
          "environment",
          "owner",
          "tags",
          "notes",
          "state",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "hostname": {
            "type": "string"
          },
          "primaryIpAddress": {
            "type": "string"
          },
          "vendor": {
            "$ref": "#/components/schemas/DeviceVendor"
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "osVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "serialNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/DeviceRole"
          },
          "criticality": {
            "$ref": "#/components/schemas/CriticalityTier"
          },
          "environment": {
            "$ref": "#/components/schemas/DeviceEnvironment"
          },
          "owner": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceEnvironment": {
        "type": "integer"
      },
      "DeviceRole": {
        "type": "integer"
      },
      "DeviceState": {
        "type": "integer"
      },
      "DeviceSummary": {
        "required": [
          "id",
          "hostname",
          "primaryIpAddress",
          "vendor",
          "model",
          "role",
          "site",
          "criticality",
          "environment",
          "state",
          "tags",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "hostname": {
            "type": "string"
          },
          "primaryIpAddress": {
            "type": "string"
          },
          "vendor": {
            "$ref": "#/components/schemas/DeviceVendor"
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/DeviceRole"
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "criticality": {
            "$ref": "#/components/schemas/CriticalityTier"
          },
          "environment": {
            "$ref": "#/components/schemas/DeviceEnvironment"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceVendor": {
        "type": "integer"
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateDeviceRequest": {
        "required": [
          "hostname",
          "primaryIpAddress"
        ],
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          },
          "primaryIpAddress": {
            "type": "string"
          },
          "vendor": {
            "$ref": "#/components/schemas/DeviceVendor"
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "osVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "serialNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/DeviceRole"
          },
          "criticality": {
            "$ref": "#/components/schemas/CriticalityTier"
          },
          "environment": {
            "$ref": "#/components/schemas/DeviceEnvironment"
          },
          "owner": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}